input style props
