+++ docs/src/pages/components/passwordfield/examples/ErrorExample.tsx
@@ -1,0 +1,10 @@
+import { PasswordField } from '@aws-amplify/ui-react';
+
+export const ErrorExample = () => (
+  <PasswordField
+    label="Password"
+    defaultValue="1234"
+    hasError={true}
+    errorMessage="Password should not be 1234! 😱"
+  />
+);
