+++ docs/src/pages/components/passwordfield/examples/HideShowPasswordExample.tsx
@@ -1,0 +1,5 @@
+import { PasswordField } from '@aws-amplify/ui-react';
+
+export const HideShowPasswordExample = () => (
+  <PasswordField label="Password" name="password" hideShowPassword={true} />
+);
