+++ docs/src/pages/components/passwordfield/examples/InputStyleExample.tsx
@@ -1,0 +1,15 @@
+import { PasswordField, useTheme } from '@aws-amplify/ui-react';
+
+export const InputStyleExample = () => {
+  const { tokens } = useTheme();
+
+  return (
+    <PasswordField
+      label="Name"
+      inputStyles={{
+        backgroundColor: tokens.colors.background.warning,
+        padding: tokens.space.large,
+      }}
+    />
+  );
+};
