+++ docs/src/pages/components/phonenumberfield/examples/AccessibilityExample.tsx
@@ -1,0 +1,12 @@
+import { Flex, PhoneNumberField } from '@aws-amplify/ui-react';
+
+export const AccessibilityExample = () => (
+  <Flex direction="column">
+    <PhoneNumberField
+      label="Phone Number"
+      labelHidden={true}
+      defaultCountryCode="+1"
+      placeholder="Phone Number"
+    />
+  </Flex>
+);
