+++ docs/src/pages/components/phonenumberfield/examples/DescriptiveTextExample.tsx
@@ -1,0 +1,20 @@
+import { View, PhoneNumberField, Text } from '@aws-amplify/ui-react';
+
+export const DescriptiveTextExample = () => (
+  <View width="100%">
+    <PhoneNumberField
+      defaultCountryCode="+1"
+      label="Phone Number"
+      descriptiveText={
+        <Text
+          as="span"
+          color="rebeccapurple"
+          fontStyle="italic"
+          fontSize="0.8rem"
+        >
+          Please enter your phone number
+        </Text>
+      }
+    />
+  </View>
+);
