+++ docs/src/pages/components/searchfield/examples/StylePropsExample.tsx
@@ -1,15 +1,14 @@
 import * as React from 'react';
 import { SearchField, useTheme } from '@aws-amplify/ui-react';
 
-export const SearchFieldStyledPropsExample = () => {
+export const StylePropsExample = () => {
   const { tokens } = useTheme();
 
   return (
     <SearchField
       label="search"
-      fontSize={tokens.fontSizes.xl}
-      backgroundColor={tokens.colors.blue[20]}
-      color={tokens.colors.black}
+      padding={tokens.space.medium}
+      backgroundColor={tokens.colors.brand.primary[10]}
     />
   );
 };
