+++ docs/src/pages/components/searchfield/examples/index.ts
@@ -1,6 +1,7 @@
 export { DefaultSearchFieldExample } from './DefaultSearchFieldExample';
 export { PlaceholderSearchFieldExample } from './PlaceholderSearchFieldExample';
-export { RefExample } from './refs';
-export { SearchFieldStyledPropsExample } from './SearchFieldStyledPropsExample';
+export { RefExample } from './RefExample';
+export { StylePropsExample } from './StylePropsExample';
 export { SizeSearchFieldExample } from './SizeSearchFieldExample';
 export { VariationSearchFieldExample } from './VariationSearchFieldExample';
+export { InputStyleExample } from './InputStyleExample';
