+++ docs/src/pages/components/shared/inputStyles.mdx
@@ -1,0 +1,1 @@
+You can set styling props on the input element by using the `inputStyles` prop and passing an object of styling props to it. The style props in `inputStyles` work the same way as style props on the component itself, you can even use [responsive props](/theming/responsive)
