+++ docs/src/pages/components/stepperfield/examples/InputStyleExample.tsx
@@ -1,0 +1,19 @@
+import { StepperField, useTheme } from '@aws-amplify/ui-react';
+
+export const InputStyleExample = () => {
+  const { tokens } = useTheme();
+  return (
+    <StepperField
+      label="Stepper"
+      defaultValue={0}
+      min={0}
+      max={10}
+      step={1}
+      inputStyles={{
+        fontWeight: tokens.fontWeights.extrabold,
+        backgroundColor: tokens.colors.background.disabled,
+      }}
+      labelHidden
+    />
+  );
+};
