+++ docs/src/pages/components/stepperfield/examples/StylePropsExample.tsx
@@ -1,6 +1,6 @@
 import { StepperField, useTheme } from '@aws-amplify/ui-react';
 
-export const StepperFieldStylePropsExample = () => {
+export const StylePropsExample = () => {
   const { tokens } = useTheme();
   return (
     <StepperField
@@ -9,8 +9,9 @@ export const StepperFieldStylePropsExample = () => {
       min={0}
       max={10}
       step={1}
-      backgroundColor={tokens.colors.teal[60]}
-      border={`1px solid ${tokens.colors.teal[60]}`}
+      fontWeight={tokens.fontWeights.extrabold}
+      padding={tokens.space.large}
+      border={`${tokens.borderWidths.large} solid ${tokens.colors.teal[80]}`}
       color={tokens.colors.white}
       labelHidden
     />
