+++ docs/src/pages/components/stepperfield/react.mdx
@@ -9,6 +9,8 @@ import {
   StepperFieldAccessibilityExample,
   StepperFieldSizeExample,
   StepperFieldValidationErrorExample,
+  StepperFieldStylePropsExample,
+  InputStyleExample,
 } from './examples';
 import { Example, ExampleCode } from '@/components/Example';
 import { Fragment } from '@/components/Fragment';
@@ -151,7 +153,7 @@ To override styling on all StepperFields, you can set the Amplify CSS variables 
 
 To override styling on a specific StepperField, you can use (in order of increasing specificity): a class selector, data attributes, or style props.
 
-_Using a class selector:_
+#### Using a class selector
 
 <Example>
   <StepperField
@@ -193,62 +195,20 @@ _Using a class selector:_
   </ExampleCode>
 </Example>
 
-_Using data attributes:_
+#### Using style props
+
+All style props will be applied to the outermost [Flex](/components/flex) component of the StepperField.
 
 <Example>
-  <StepperField
-    label="Stepper"
-    className="data-attribute-large-size"
-    defaultValue={0}
-    min={0}
-    max={10}
-    size="large"
-    step={1}
-    labelHidden
-  />
+  <StepperFieldStylePropsExample />
   <ExampleCode>
-    ```css
-    /* styles.css */
-
-    /* Override only large size styles */
-    .amplify-stepperfield[data-size='large'] {
-      width: 50%;
-    }
-    ```
-
+    ```jsx file=./examples/StepperFieldStylePropsExample.tsx ```
   </ExampleCode>
 </Example>
 
-_Using style props:_
+<Fragment>{() => import('./../shared/inputStyles.mdx')}</Fragment>
 
 <Example>
-  <StepperField
-    label="Stepper"
-    defaultValue={0}
-    min={0}
-    max={10}
-    step={1}
-    backgroundColor="var(--amplify-colors-teal-60)"
-    border="1px solid var(--amplify-colors-teal-60)"
-    color="var(--amplify-colors-white)"
-    labelHidden
-  />
-  <ExampleCode>
-    ```jsx
-    import { StepperField } from '@aws-amplify/ui-react';
-
-    <StepperField
-      label="Stepper"
-      defaultValue={0}
-      min={0}
-      max={10}
-      step={1}
-      color="var(--amplify-colors-white)"
-      border="1px solid var(--amplify-colors-teal-60)"
-      backgroundColor="var(--amplify-colors-teal-60)"
-      labelHidden
-    />;
-    ```
-
-  </ExampleCode>
+  <InputStyleExample />
+  <ExampleCode>```jsx file=./examples/InputStyleExample.tsx ```</ExampleCode>
 </Example>
