+++ docs/src/pages/components/textfield/examples/StylePropsExample.tsx
@@ -1,6 +1,6 @@
 import { Text, TextField, useTheme } from '@aws-amplify/ui-react';
 
-export const TextFieldStylePropsExample = () => {
+export const StylePropsExample = () => {
   const { tokens } = useTheme();
   return (
     <TextField
@@ -15,8 +15,9 @@ export const TextFieldStylePropsExample = () => {
           Name:
         </Text>
       }
-      backgroundColor={tokens.colors.background.primary}
-      color={tokens.colors.black}
+      padding={tokens.space.large}
+      backgroundColor={tokens.colors.brand.primary[80]}
+      color={tokens.colors.font.inverse}
       width="400px"
     />
   );
