+++ docs/src/styles/docs/base.scss
@@ -134,8 +134,14 @@ h3 {
   font-weight: var(--amplify-components-heading-3-font-weight);
 }
 
+h4 {
+  font-size: var(--amplify-components-heading-4-font-size);
+  font-weight: var(--amplify-components-heading-4-font-weight);
+}
+
 h2[id],
-h3[id] {
+h3[id],
+h4[id] {
   position: relative;
   margin-top: var(--amplify-space-large);
   margin-bottom: var(--amplify-space-small);
@@ -158,7 +164,8 @@ h2[id] {
 }
 
 h2[id] > a,
-h3[id] > a {
+h3[id] > a,
+h4[id] > a {
   position: absolute;
   left: -1.75rem;
   text-decoration: none;
