+++ packages/react/src/primitives/types/phoneNumberField.ts
@@ -2,6 +2,7 @@ import * as React from 'react';
 
 import { SelectFieldProps } from './selectField';
 import { TextInputFieldProps } from './textField';
+import { BaseStyleProps } from './style';
 
 export interface PhoneNumberFieldProps extends TextInputFieldProps {
   countryCodeLabel?: string;
@@ -13,6 +14,11 @@ export interface PhoneNumberFieldProps extends TextInputFieldProps {
    * Forwarded ref for access to Country Code select DOM element
    */
   countryCodeRef?: React.Ref<HTMLSelectElement>;
+
+  /**
+   * Styling that is applied to the select component of the field
+   */
+  selectStyles?: BaseStyleProps;
 }
 
 export interface CountryCodeSelectProps extends SelectFieldProps {
